Obligatory fix for Win64.
BUG=v8:1832
TEST=none
TBR=vitalyr@chromium.org

Review URL: http://codereview.chromium.org/8771051

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@10183 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/profile-generator-inl.h
@@ -108,9 +108,10 @@ HeapObject* V8HeapExplorer::GetNthGcSubrootObject(int delta) {
 
 
 int V8HeapExplorer::GetGcSubrootOrder(HeapObject* subroot) {
-  return (reinterpret_cast<char*>(subroot) -
-          reinterpret_cast<char*>(kFirstGcSubrootObject)) /
-      HeapObjectsMap::kObjectIdStep;
+  return static_cast<int>(
+      (reinterpret_cast<char*>(subroot) -
+       reinterpret_cast<char*>(kFirstGcSubrootObject)) /
+      HeapObjectsMap::kObjectIdStep);
 }
